Moved nav-link fns to helper file
